refactor: 방 목록 관리 로직 커스텀 훅으로 추출
- useRooms 훅 생성하여 중복 코드 제거 (LobbyPage, RoomListPage)

- RoomItem 컴포넌트를 별도 파일로 분리하여 재사용성 향상

- 무한 재요청 문제 해결 (useEffect 의존성 최적화)

- 약 300줄 이상의 중복 코드 제거로 유지보수성 개선
